Start service after deb/rpm install or upgrade if env file exists (#805)
* Start service after deb/rpm install or upgrade if env file exists

* Add commands to check status and logs

+++ internal/buildscripts/packaging/fpm/preuninstall.sh
@@ -15,7 +15,9 @@
 # limitations under the License.
 
 if command -v systemctl >/dev/null 2>&1; then
-    echo "Stopping splunk-otel-collector service"
-    systemctl stop splunk-otel-collector.service
     systemctl disable splunk-otel-collector.service
+    if systemctl status splunk-otel-collector.service >/dev/null 2>&1; then
+        echo "Stopping splunk-otel-collector service"
+        systemctl stop splunk-otel-collector.service
+    fi
 fi
